requirements.txt
telethon==1.36.0
Pillow==10.4.0
PyYAML==6.0.2
python-dotenv==1.0.1
httpx==0.27.0
python-telegram-bot==21.3
pytz==2024.1
apscheduler==3.10.4
pytest==8.2.0
